# documentation/backend/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Alforis CRM API",
    "version": "1.0.0",
    "description": "Documentation de référence de l'API CRM Alforis. Les endpoints décrits couvrent les fonctionnalités principales livrées dans le plan d'amélioration : authentification, organisations, personnes, tâches, mandats, recherche globale, exports et webhooks.",
    "contact": {
      "name": "Équipe CRM Alforis",
      "email": "tech@alforis-finance.com"
    }
  },
  "servers": [
    {
      "url": "https://crm.alforis-finance.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
      "description": "Développement local"
    }
  ],
...
  "paths": {
...
    "/api/v1/organisations": {
      "get": {
        "tags": [
          "organisations"
        ],
        "summary": "Lister les organisations",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Offset de pagination"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000
            },
            "description": "Taille de page"
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Institution",
                "Wholesale",
                "SDG",
                "CGPI",
                "Autres"
              ]
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "country_code",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 2
            }
          },
          {
            "name": "language",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page d'organisations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedOrganisations"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "organisations"
        ],
        "summary": "Créer une organisation",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganisationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organisation créée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organisation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/organisations/search": {
      "get": {
        "tags": [
          "organisations"
        ],
        "summary": "Recherche textuelle organisations",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Résultats paginés",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedOrganisations"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/organisations/stats": {
      "get": {
        "tags": [
          "organisations"
        ],
        "summary": "Statistiques globales",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Statistiques",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationStats"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/organisations/by-language/{language}": {
      "get": {
        "tags": [
          "organisations"
        ],
        "summary": "Organisations par langue",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "language",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 5
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Résultats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedOrganisations"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/organisations/{organisation_id}": {
      "get": {
        "tags": [
          "organisations"
        ],
        "summary": "Détail d'une organisation",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "organisation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organisation détaillée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationDetail"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "put": {
        "tags": [
          "organisations"
        ],
        "summary": "Mettre à jour une organisation",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "organisation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganisationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Organisation mise à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organisation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "organisations"
        ],
        "summary": "Supprimer une organisation",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "organisation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Supprimé"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v1/mandats/organisation/{organisation_id}": {
      "get": {
        "tags": [
          "mandats"
        ],
        "summary": "Mandats d'une organisation",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "organisation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Mandat"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Recherche globale multi-entités",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2
            }
          },
          {
            "name": "types",
            "in": "query",
            "schema": {
              "type": "string",
              "description": "organisations,people,mandats,tasks"
            }
          },
          {
            "name": "limit_per_type",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Résultats regroupés",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchGlobalResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/search/organisations": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Recherche avancée organisations",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Institution",
                "Wholesale",
                "SDG",
                "CGPI",
                "Autres"
              ]
            }
          },
          {
            "name": "city",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "pipeline_stage",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Résultats paginés",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchCollection"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Organisation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "example": "ACME Capital"
          },
          "category": {
            "type": "string",
            "enum": [
              "Institution",
              "Wholesale",
              "SDG",
              "CGPI",
              "Autres"
            ]
          },
          "aum": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "aum_date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "strategies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "website": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "country_code": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "nullable": true
          },
          "domicile": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "minLength": 2,
            "maxLength": 5,
            "example": "FR"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "is_active": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "language",
          "is_active"
        ]
      },
      "OrganisationCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "Institution",
              "Wholesale",
              "SDG",
              "CGPI",
              "Autres"
            ]
          },
          "aum": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "aum_date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "strategies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "website": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "country_code": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "nullable": true
          },
          "domicile": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "minLength": 2,
            "maxLength": 5,
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "is_active": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "name",
          "category"
        ]
      },
      "OrganisationUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "aum": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "aum_date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "strategies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "website": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "country_code": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "nullable": true
          },
          "domicile": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "minLength": 2,
            "maxLength": 5,
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "is_active": {
            "type": "boolean",
            "nullable": true
          }
        }
      },
      "OrganisationDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Organisation"
          },
          {
            "type": "object",
            "properties": {
              "mandats": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Mandat"
                }
              },
              "contacts": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "organisation_id": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string",
                      "format": "email",
                      "nullable": true
                    },
                    "phone": {
                      "type": "string",
                      "nullable": true
                    },
                    "title": {
                      "type": "string",
                      "nullable": true
                    },
                    "notes": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "organisation_id",
                    "name"
                  ]
                }
              }
            }
          }
        ]
      },
      "OrganisationStats": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "by_category": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "by_language": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          }
        },
        "required": [
          "total",
          "by_category",
          "by_language"
        ]
      },
      "PaginatedOrganisations": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "skip": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Organisation"
            }
          }
        },
        "required": [
          "total",
          "skip",
          "limit",
          "items"
        ]
      },
...
      "Mandat": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "organisation_id": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "proposé",
              "signé",
              "actif",
              "terminé"
            ]
          },
          "date_signature": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "date_debut": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "date_fin": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "is_actif": {
            "type": "boolean",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "organisation_id",
          "status"
        ]
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Auth requise",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Erreur de validation",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "loc": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "msg": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Ressource introuvable",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}